requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
idna==3.10
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
